Config generation and deployment instructions (#5)
* Updating bootstrap.sh to apply overlay to existing config files for flexible reconfiguration

* Chart baseline

* Fixing issue with cleanup in bootstrap

* Config generation script WIP

* Fleshed out nsc config

* General improvments. Added nsc table json parser

* Added missing operator sk edit command. Fixed path typo. Added file write prompt

* Adding dependency check. Fixing issue with manual path prompt for creds and nkey

* Added advanced and helm flags. Added functionality to output base64 encoded secrets file for helm chart to consume. Various fixes.

* Fixing issue with bools written as strings

* Updated secret keys. Fixed issue with helm secrets flow

* Updating config generation with helm flag and registry credential prompt. Fixed bugs in nsc table parser. Improvements with nsc env handling.

* Fixed macOS/Linux compat issues

* Removing helm chart. Moving demo deployment to 'demo' directory. Paring down docker-compose.

* Updating readme. Minor bugfix/cleanup

* Minor README tweaks

+++ demo/bootstrap.sh
@@ -4,6 +4,7 @@ cd "/conf"
 
 OVERLAY_CONFIG_FILE="helix.json"
 RNA_CONFIG_FILE="rna.cue"
+NEW_INSTALL="true"
 
 CLUSTERS="nats-a nats-b nats-c"
 CONFIG_DIR="/conf/helix"
@@ -36,12 +37,44 @@ RNA_CONFIG=$(cat <<EOF
 EOF
 )
 
+overlay () {
+  local RNA_CONFIG="$1"
+
+  OVERLAY_CONFIG=$(cat ${SETUP_DIR}/${OVERLAY_CONFIG_FILE} | jq -e 'if . == {} then null else . end')
+
+  if [[ ${?} -gt 1 ]]; then
+    echo "Invalid ${OVERLAY_CONFIG_FILE}" >&2
+    cleanup
+    exit 1
+  elif [[ "${OVERLAY_CONFIG}" != "null" ]]; then
+    echo -e "Overlaying ${OVERLAY_CONFIG_FILE}" >&2
+    OUT_CONFIG=$(echo -e "${RNA_CONFIG}" "${OVERLAY_CONFIG}" | jq -s '.[0] * .[1]')
+    if [[ ${?} -ne 0 ]]; then
+      echo "Overlay Failed" >&2
+      OUT_CONFIG="${RNA_CONFIG}"
+    fi
+    echo "${OUT_CONFIG}"
+  fi
+}
+
+write_config () {
+  echo -e "${RNA_CONFIG}" | jq > "${CONFIG_DIR}/${RNA_CONFIG_FILE}"
+}
+
 cleanup () {
-  rm -rf ${CONFIG_DIR}
+  if [[ ${NEW_INSTALL} == "true" ]]; then
+    rm -rf ${CONFIG_DIR}
+  fi
 }
 
 if [[ -d ${CONFIG_DIR} ]] && [[ "${1}" == "-s" ]]; then
   echo "Using existing config directory"
+  NEW_INSTALL="false"
+  # Apply overlay if present
+  if [[ -f ${SETUP_DIR}/${OVERLAY_CONFIG_FILE} ]]; then
+    RNA_CONFIG=$(overlay "$(cat "${CONFIG_DIR}/${RNA_CONFIG_FILE}")")
+    write_config
+  fi
   exit 0
 elif [[ -d ${CONFIG_DIR} ]]; then
   if [[ "${1}" != "-f" ]]; then
@@ -105,21 +138,13 @@ done
 # Append nats system array to RNA config
 RNA_CONFIG="${RNA_CONFIG}${NATS_SYSTEMS::-1}]\n}"
 
+# Apply overlay if present
 if [[ -f ${SETUP_DIR}/${OVERLAY_CONFIG_FILE} ]]; then
-  OVERLAY_CONFIG=$(cat ${SETUP_DIR}/${OVERLAY_CONFIG_FILE} | jq -e 'if . == {} then null else . end')
-  
-  if [[ ${?} -gt 1 ]]; then
-    echo "Invalid ${OVERLAY_CONFIG_FILE}"
-    cleanup
-    exit 1
-  elif [[ "${OVERLAY_CONFIG}" != "null" ]]; then
-    echo -e "Overlaying ${OVERLAY_CONFIG_FILE}\n${OVERLAY_CONFIG}"
-    RNA_CONFIG=$(echo -e "${RNA_CONFIG}" "${OVERLAY_CONFIG}" | jq -s '.[0] * .[1]')
-  fi
+  RNA_CONFIG=$(overlay "${RNA_CONFIG}")
 fi
 
 # Write out RNA config to file
-echo -e "${RNA_CONFIG}" | jq > "${CONFIG_DIR}/${RNA_CONFIG_FILE}"
+write_config
 
 # Fail safely if invalid JSON
 if [[ ${?} -ne 0 ]]; then
